requirements.txt

Django==2.2.5
django-bootstrap4==2.3.1
django-clear-cache==0.3
django-debug-toolbar==3.1.1
django-phone-field==1.8.1
docutils==0.16
entrypoints==0.3

psycopg2==2.8.6
